Extract interfaces for Schedulers, add javadoc, add factory methods
